made some changes to add .docx in Puzzle skills.
