+++ Gemfile
@@ -48,6 +48,7 @@ gem 'rails-i18n'
 gem 'rest-client'
 gem 'rswag-api'
 gem 'rswag-ui'
+gem 'sablon'
 gem 'seed-fu'
 gem 'sentry-rails'
 gem 'sentry-ruby'
