+++ spec/exporters/odt/cv_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
-describe Odt::Cv do
+describe Export::BaseCvExport do
   fixtures :people
 
-  before { allow_any_instance_of(Odt::Cv).to receive(:location).and_return(branch_adresses(:bern)) }
+  before { allow_any_instance_of(Export::BaseCvExport).to receive(:location).and_return(branch_adresses(:bern)) }
 
   let(:person) { double('Person') }
   let(:params) do
@@ -20,49 +20,49 @@ describe Odt::Cv do
       person = people(:bob)
       person.remove_picture
       expect(person.picture.file).to be_nil
-      Odt::Cv.new(person, { 'anon' => 'false' })
+      Export::BaseCvExport.new(person, { 'anon' => 'false' })
     end
 
     it 'can export without competence_notes' do
       person = people(:bob)
       person.competence_notes = nil
-      Odt::Cv.new(person, { 'anon' => 'false' })
+      Export::BaseCvExport.new(person, { 'anon' => 'false' })
     end
 
     it 'translates nationalities' do
-      nationalities = Odt::Cv.new(people(:bob), { 'anon' => 'false' }).send(:nationalities)
+      nationalities = Export::BaseCvExport.new(people(:bob), { 'anon' => 'false' }).send(:nationalities)
       expect(nationalities).to eq('Schweiz, Schweden')
 
-      nationalities = Odt::Cv.new(people(:alice), { 'anon' => 'false' }).send(:nationalities)
+      nationalities = Export::BaseCvExport.new(people(:alice), { 'anon' => 'false' }).send(:nationalities)
       expect(nationalities).to eq('Australien')
     end
 
     it 'get one skill hash when filter on level 3' do
-      skills = Odt::Cv.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level_value, 3)
+      skills = Export::BaseCvExport.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level_value, 3)
 
       expect(skills).to eq([{ :category => "Software-Engineering",
                               :competence => "Kompetenzen des Levels Professional: \r\n- Rails \r\n" }])
     end
 
     it 'get empty skill hash when filter on level 5' do
-      skills = Odt::Cv.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level_value, 5)
+      skills = Export::BaseCvExport.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level_value, 5)
 
       expect(skills).to eq([])
     end
 
     it 'get PeopleSkill by level 2' do
-      skills = Odt::Cv.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level, 2)
+      skills = Export::BaseCvExport.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level, 2)
       people_skill = PeopleSkill.find 996949822
       expect(skills).to eq([people_skill])
     end
 
     it 'get no PeopleSkill by level 5' do
-      skills = Odt::Cv.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level, 5)
+      skills = Export::BaseCvExport.new(people(:bob), { 'skillsByLevel' => 'true' }).send(:skills_by_level, 5)
       expect(skills).to eq([])
     end
 
     it 'formats competence notes' do
-      notes = Odt::Cv.new(people(:bob), { 'anon' => 'false' }).send(:competence_notes_list)[:competence]
+      notes = Export::BaseCvExport.new(people(:bob), { 'anon' => 'false' }).send(:competence_notes_list)[:competence]
       expect(notes).to eq("Java\nRuby")
     end
 
@@ -84,7 +84,7 @@ describe Odt::Cv do
 
   describe 'format language with level correctly' do
     let(:bob) { people(:bob) }
-    let(:cv) { Odt::Cv.new(bob, { 'anon' => 'false' }) }
+    let(:cv) { Export::BaseCvExport.new(bob, { 'anon' => 'false' }) }
     let(:report) { instance_double('ODFReport::Report') }
     let(:german) { language_skills(:deutsch)}
